Preprocessing/training
First I built a train and test split on ticket prices where we used 70% of the numerical data for training and 30% for testing. I tested the training data to ensure it was a good representation of the overall data structure by testing it against average price. 

I then used a linear regression model to test if it was better to fill the missing values with mean or median finding that each value returned about the same. I dove deeper into this by creating a pipeline to easily impute and scale the data and perform linear regression. After this I tested and refined the pipeline with SelectKBest then performed cross vaildation to see what would be the best value for k based on variability of our data. With k=15 I found that there was higher variability than we would have liked. When plotting the values of K vs variability, I found that 8 was the best k value as further values may be overfit. I found that with the linera regression, verical drop was the biggest positive feature for ticket prices and things such as skiable terrain and trams were negative with the linear regression model.

I also built a random forest model to compare how it performed against the linear regression model. I used a pipeline and cross validation similar to the linear regression model, finding that our cross-validation results marginally improved. Using the random forest model, I found that fast quads and runs had the largest impact on ticket price while trams had a very low impact

After comparing model performance, I decided to move forward with the Random Forest regression model because is had a lower mean abosolute error during our cross-validation signaling less variability.

I then verified we were using enough data to ensure that we did not need further data collection. I found that I had plenty of data and after about 60 data points, the CV scores completely leveled out.